requirements.txt
Biopython
matplotlib